requirements.txt
# Core
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0

# 3D object source
objaverse>=0.1.7
